+++ src/components/file/examples/file.tsx
@@ -7,7 +7,7 @@ import { Component, h, State } from '@stencil/core';
 })
 export class FileExample {
     @State()
-    private value: FileInfo = { filename: 'bla.jpg', id: 123 };
+    private value: FileInfo = { filename: 'picture.jpg', id: 123 };
 
     @State()
     private required = false;
@@ -19,17 +19,19 @@ export class FileExample {
 
     public render() {
         return [
-            <limel-switch
-                label="Toggle required"
-                value={this.required}
-                onChange={this.toggleRequired}
-            />,
             <limel-file
-                label="File"
+                label="Attach a file"
                 value={this.value}
                 required={this.required}
                 onChange={this.handleChange}
             />,
+            <limel-flex-container justify="end">
+                <limel-switch
+                    label="Toggle required"
+                    value={this.required}
+                    onChange={this.toggleRequired}
+                />
+            </limel-flex-container>,
         ];
     }