+++ src/components/form/examples/custom-component-form.tsx
@@ -48,10 +48,7 @@ export class CustomComponentFormExample {
                 value={this.formData}
                 schema={schema}
             />,
-            <br />,
-            <br />,
-            'Value: ',
-            <pre>{JSON.stringify(this.formData, null, '    ')}</pre>,
+            <limel-example-value value={this.formData} />,
         ];
     }