+++ src/components/input-field/examples/input-field-autocomplete.tsx
@@ -61,7 +61,7 @@ export class InputFieldAutocompleteExample {
                     />
                 </limel-flex-container>
             </p>,
-            <p>Value: {this.value}</p>,
+            <limel-example-value value={this.value} />,
         ];
     }