+++ src/components/input-field/examples/input-field-number.tsx
@@ -60,7 +60,7 @@ export class InputFieldNumberExample {
                     />
                 </limel-flex-container>
             </p>,
-            <p>Value: {this.value}</p>,
+            <limel-example-value value={this.value} />,
         ];
     }