+++ src/components/input-field/examples/input-field-textarea.tsx
@@ -3,7 +3,6 @@ import { Component, h, State } from '@stencil/core';
 @Component({
     tag: 'limel-example-input-field-textarea',
     shadow: true,
-    styleUrl: 'input-field.scss',
 })
 export class InputFieldTextareaExample {
     @State()
@@ -17,8 +16,8 @@ export class InputFieldTextareaExample {
 
     constructor() {
         this.changeHandler = this.changeHandler.bind(this);
-        this.toggleEnabled = this.toggleEnabled.bind(this);
-        this.toggleRequired = this.toggleRequired.bind(this);
+        this.setEnabled = this.setEnabled.bind(this);
+        this.setRequired = this.setRequired.bind(this);
     }
 
     public render() {
@@ -38,16 +37,16 @@ export class InputFieldTextareaExample {
             <p>
                 <limel-flex-container justify="end">
                     <limel-checkbox
-                        onClick={this.toggleEnabled}
+                        onChange={this.setEnabled}
                         label="Enabled"
                     />
                     <limel-checkbox
-                        onClick={this.toggleRequired}
+                        onChange={this.setRequired}
                         label="Required"
                     />
                 </limel-flex-container>
             </p>,
-            <p>Value: {this.value}</p>,
+            <limel-example-value value={this.value} />,
         ];
     }
 
@@ -55,11 +54,13 @@ export class InputFieldTextareaExample {
         this.value = event.detail;
     }
 
-    private toggleEnabled() {
-        this.disabled = !this.disabled;
+    private setEnabled(event: CustomEvent<boolean>) {
+        event.stopPropagation();
+        this.disabled = event.detail;
     }
 
-    private toggleRequired() {
-        this.required = !this.required;
+    private setRequired(event: CustomEvent<boolean>) {
+        event.stopPropagation();
+        this.required = event.detail;
     }
 }