+++ src/components/list/examples/list-checkbox-icons.tsx
@@ -4,7 +4,6 @@ import { Component, h, State } from '@stencil/core';
 @Component({
     tag: 'limel-example-list-checkbox-icons',
     shadow: true,
-    styleUrl: 'list.scss',
 })
 export class ListCheckboxIconsExample {
     @State()
@@ -72,9 +71,7 @@ export class ListCheckboxIconsExample {
                 items={this.items}
                 type="checkbox"
             />,
-            <p>
-                Value: <code>{JSON.stringify(this.selectedItems)}</code>
-            </p>,
+            <limel-example-value value={this.selectedItems} />,
         ];
     }