+++ src/components/list/examples/list-radio-button-icons.tsx
@@ -4,7 +4,6 @@ import { Component, h, State } from '@stencil/core';
 @Component({
     tag: 'limel-example-list-radio-button-icons',
     shadow: true,
-    styleUrl: 'list.scss',
 })
 export class ListRadioButtonIconsExample {
     @State()
@@ -72,9 +71,7 @@ export class ListRadioButtonIconsExample {
                 items={this.items}
                 type="radio"
             />,
-            <p>
-                Value: <code>{JSON.stringify(this.selectedItem)}</code>
-            </p>,
+            <limel-example-value value={this.selectedItem} />,
         ];
     }