+++ src/components/list/examples/list-radio-button.tsx
@@ -4,7 +4,6 @@ import { Component, h, State } from '@stencil/core';
 @Component({
     tag: 'limel-example-list-radio-button',
     shadow: true,
-    styleUrl: 'list.scss',
 })
 export class ListRadioButtonExample {
     @State()
@@ -35,9 +34,7 @@ export class ListRadioButtonExample {
                 items={this.items}
                 type="radio"
             />,
-            <p>
-                Value: <code>{JSON.stringify(this.selectedItem)}</code>
-            </p>,
+            <limel-example-value value={this.selectedItem} />,
         ];
     }