+++ src/components/tab-bar/examples/tab-bar-with-dynamic-tab-width.tsx
@@ -39,7 +39,7 @@ export class TabBarExample {
     public render() {
         return [
             <limel-tab-bar tabs={this.tabs} onChangeTab={this.onChange} />,
-            <p>Tab: {this.text}</p>,
+            <limel-example-value label="Tab" value={this.text} />,
         ];
     }