+++ src/components/table/examples/table-activate-row.tsx
@@ -35,11 +35,9 @@ export class TableExampleActivateRow {
                 activeRow={this.activeRow}
                 columns={this.columns}
                 onActivate={this.onActivateRow}
-                class="has-low-density has-interactive-rows"
+                class="has-interactive-rows"
             />,
-            <p class="active-row">
-                Active row: <code>{JSON.stringify(this.activeRow)}</code>
-            </p>,
+            <limel-example-value label="Active row" value={this.activeRow} />,
         ];
     }